[refactor]: rename folder content->post
