+++ plzdrawing/src/screens/post/HomePostDetail.tsx
